+++ libpod/networking_freebsd.go
@@ -79,39 +79,6 @@ type LinkStatistics64 struct {
 	TxCompressed      uint64
 }
 
-// convertPortMappings will remove the HostIP part from the ports when running inside podman machine.
-// This is need because a HostIP of 127.0.0.1 would now allow the gvproxy forwarder to reach to open ports.
-// For machine the HostIP must only be used by gvproxy and never in the VM.
-func (c *Container) convertPortMappings() []types.PortMapping {
-	if !c.runtime.config.Engine.MachineEnabled || len(c.config.PortMappings) == 0 {
-		return c.config.PortMappings
-	}
-	// if we run in a machine VM we have to ignore the host IP part
-	newPorts := make([]types.PortMapping, 0, len(c.config.PortMappings))
-	for _, port := range c.config.PortMappings {
-		port.HostIP = ""
-		newPorts = append(newPorts, port)
-	}
-	return newPorts
-}
-
-func (c *Container) getNetworkOptions(networkOpts map[string]types.PerNetworkOptions) types.NetworkOptions {
-	opts := types.NetworkOptions{
-		ContainerID:   c.config.ID,
-		ContainerName: getCNIPodName(c),
-	}
-	opts.PortMappings = c.convertPortMappings()
-
-	// If the container requested special network options use this instead of the config.
-	// This is the case for container restore or network reload.
-	if c.perNetworkOpts != nil {
-		opts.Networks = c.perNetworkOpts
-	} else {
-		opts.Networks = networkOpts
-	}
-	return opts
-}
-
 type RootlessNetNS struct {
 	dir  string
 	Lock lockfile.Locker
